moved simtime into state object. moved loader functions in their own files. removed the unused dashboardstate type
