+++ gameModel/libs/ClientScript/game/common/simulationState/loaders/taskLoader.ts
@@ -1,9 +1,13 @@
 import { LOCATION_ENUM } from '../../simulationState/locationState';
-import { HealingTask, TaskBase } from '../../tasks/taskBase';
+import { HealingTask, TaskBase, TaskType } from '../../tasks/taskBase';
 import { EvacuationTask } from '../../tasks/taskBaseEvacuation';
 import { PorterTask } from '../../tasks/taskBasePorter';
 import { PreTriageTask } from '../../tasks/taskBasePretriage';
 import { WaitingTask } from '../../tasks/taskBaseWaiting';
+
+export function getWaitingTaskId(loadedTasks: TaskBase[]): number {
+  return loadedTasks.find(t => t.taskType === TaskType.Waiting)!.Uid;
+}
 
 export function loadTasks(): TaskBase[] {
   const taskPretriChantier = new PreTriageTask(
